+++ src/main/java/dev/nyon/bbm/asm/AbstractBoatMixin.java
@@ -6,11 +6,11 @@ import dev.nyon.bbm.logic.BbmBoat;
 import dev.nyon.bbm.config.Config;
 import dev.nyon.bbm.config.ConfigKt;
 import net.minecraft.core.BlockPos;
-import net.minecraft.tags.BlockTags;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.player.Player;
 //? if >=1.21.3
 import net.minecraft.world.entity.vehicle.AbstractBoat;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
@@ -24,6 +24,7 @@ import org.spongepowered.asm.mixin.injection.ModifyConstant;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 @SuppressWarnings("AddedMixinMembersNamePattern")
 @Pseudo
@@ -32,6 +33,9 @@ public class AbstractBoatMixin implements BbmBoat {
 
     @Unique
     private boolean jumpCollision = false;
+
+    @Unique
+    private boolean correctCollision = false;
 
     @Override
     public void setJumpCollision(boolean b) {
@@ -43,13 +47,18 @@ public class AbstractBoatMixin implements BbmBoat {
         return jumpCollision;
     }
 
+    @Override
+    public void setCorrectCollision(boolean b) { correctCollision = b; }
+
+    @Override
+    public boolean getCorrectCollision() { return correctCollision; }
+
     /*? if >=1.21.3 {*/
     @Shadow
     private AbstractBoat.Status status;
     @Unique
     private AbstractBoat instance = (AbstractBoat) (Object) this;
 
-    @SuppressWarnings("resource")
     @Unique
     private List<BlockState> getCarryingBlocks() {
         List<BlockState> states = new ArrayList<>();
@@ -97,45 +106,20 @@ public class AbstractBoatMixin implements BbmBoat {
         BbmBoat bbmBoat = (BbmBoat) instance;
 
         if (!config.getBoosting().getBoostStates().contains(status)) return original;
+        if (!ConfigCacheKt.getAllowedCollidingBlocks().isEmpty() && !bbmBoat.getCorrectCollision()) return original;
         if (status == AbstractBoat.Status.ON_LAND) {
-            if (!ConfigCacheKt.getAllowedSupportingBlocks().isEmpty()) {
+            Set<Block> allowedBlocks = ConfigCacheKt.getAllowedSupportingBlocks();
+            if (!allowedBlocks.isEmpty()) {
                 List<BlockState> carryingBlocks = getCarryingBlocks();
-                if (carryingBlocks.stream().noneMatch(state -> ConfigCacheKt.getAllowedSupportingBlocks().contains(state.getBlock()))
+                if (carryingBlocks.stream().noneMatch(state -> allowedBlocks.contains(state.getBlock())))
                     return original;
             }
-            if (!ConfigCacheKt.getAllowedCollidingBlocks().isEmpty()) {
-
-            }
         }
 
-        switch (status) {
-            case ON_LAND -> {
-                if (!ConfigKt.getActiveConfig()
-                    .getBoostOnBlocks() && !ConfigKt.getActiveConfig()
-                    .getBoostOnIce()) return original;
-                if (ConfigKt.getActiveConfig()
-                    .getBoostOnIce()) {
-                    List<BlockState> carryingBlocks = getCarryingBlocks();
-                    if (carryingBlocks.stream()
-                        .noneMatch(state -> state.is(BlockTags.ICE))) return original;
-                }
-                if (!bbmBoat.getJumpCollision() && !instance.horizontalCollision) return original;
-            }
-            case IN_WATER -> {
-                if (!ConfigKt.getActiveConfig()
-                    .getBoostOnWater()) return original;
-                if (!bbmBoat.getJumpCollision() && !instance.horizontalCollision) return original;
-            }
-            case UNDER_WATER, UNDER_FLOWING_WATER -> {
-                if (!ConfigKt.getActiveConfig()
-                    .getBoostUnderwater()) return original;
-            }
-            case IN_AIR -> {
-                return original;
-            }
-        }
+        if (!bbmBoat.getJumpCollision() && !instance.horizontalCollision) return original;
 
         bbmBoat.setJumpCollision(false);
+        bbmBoat.setCorrectCollision(false);
         return new Vec3(
             original.x,
             ConfigKt.getActiveConfig()