+++ src/main/java/dev/nyon/bbm/asm/EntityMixin.java
@@ -1,11 +1,14 @@
 package dev.nyon.bbm.asm;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import dev.nyon.bbm.config.ConfigCacheKt;
 import dev.nyon.bbm.extensions.NetworkingKt;
 import dev.nyon.bbm.logic.BbmBoat;
 import dev.nyon.bbm.config.Config;
 import dev.nyon.bbm.config.ConfigKt;
 import dev.nyon.bbm.logic.JumpCollisionPacket;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.Entity;
@@ -16,21 +19,35 @@ import net.minecraft.world.entity.vehicle.AbstractBoat;
 //? if <1.21.3
 /*import net.minecraft.world.entity.vehicle.Boat;*/
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.VoxelShape;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.Redirect;
 
 import java.util.List;
+import java.util.Set;
 
 @Mixin(Entity.class)
 public abstract class EntityMixin {
     @Shadow
     protected abstract Vec3 collide(Vec3 movement);
+
+    @Shadow
+    private static Vec3 collideWithShapes(
+        Vec3 movement,
+        AABB entityBoundingBox,
+        List<VoxelShape> shapes
+    ) {
+        return null;
+    }
 
     @Unique
     private boolean expandCollision = false;
@@ -59,13 +76,56 @@ public abstract class EntityMixin {
         Config config = ConfigKt.getActiveConfig();
         if (config == null) return box;
 
-        if (boat.getPassengers()
-            .stream()
-            .filter(passenger -> passenger instanceof Player)
-            .toList()
-            .isEmpty() && config.getBoosting().getOnlyForPlayers()) return box;
+        if (
+            config.getBoosting().getOnlyForPlayers()
+                && boat.getPassengers()
+                .stream()
+                .filter(passenger -> passenger instanceof Player)
+                .toList()
+                .isEmpty()
+        ) return box;
 
-        return box.inflate(config.getBoosting().getExtraCollisionDetectionRange(), 0, config.getBoosting().getExtraCollisionDetectionRange());
+        return box.inflate(
+            config.getBoosting().getExtraCollisionDetectionRange(),
+            0,
+            config.getBoosting().getExtraCollisionDetectionRange()
+        );
+    }
+
+    @Redirect(
+        method = "collideBoundingBox",
+        at = @At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/world/entity/Entity;collideWithShapes(Lnet/minecraft/world/phys/Vec3;Lnet/minecraft/world/phys/AABB;Ljava/util/List;)Lnet/minecraft/world/phys/Vec3;"
+        )
+    )
+    private static Vec3 checkForHorizontalCollision(
+        Vec3 vec3,
+        AABB bB,
+        List<VoxelShape> foundCollisions,
+        @Nullable Entity entity,
+        Vec3 movement,
+        AABB entityBoundingBox,
+        Level world,
+        List<VoxelShape> expectedCollisions
+    ) {
+        if (!(entity instanceof BbmBoat bbmBoat)) return collideWithShapes(vec3, bB, foundCollisions);
+
+        Set<Block> allowedSupportingBlocks = ConfigCacheKt.getAllowedCollidingBlocks();
+        if (!allowedSupportingBlocks.isEmpty()) {
+            boolean correctCollision = foundCollisions.stream().anyMatch(shape -> {
+                BlockPos blockPos = new BlockPos(
+                    shape.getCoords(Direction.Axis.X).getFirst().intValue(),
+                    shape.getCoords(Direction.Axis.Y).getFirst().intValue(),
+                    shape.getCoords(Direction.Axis.Z).getFirst().intValue()
+                );
+                BlockState blockState = world.getBlockState(blockPos);
+                return allowedSupportingBlocks.contains(blockState.getBlock());
+            });
+
+            bbmBoat.setCorrectCollision(correctCollision);
+        }
+        return collideWithShapes(vec3, bB, foundCollisions);
     }
 
     @ModifyExpressionValue(
@@ -89,7 +149,8 @@ public abstract class EntityMixin {
         boolean xEqual = Mth.equal(movement.x, withFakeBb.x);
         boolean zEqual = Mth.equal(movement.z, withFakeBb.z);
         if (xEqual && zEqual) return original;
-        // Set the jump collision for server executed entities that may control the boat
+        if (!ConfigCacheKt.getAllowedCollidingBlocks().isEmpty() && !bbmBoat.getCorrectCollision()) return original;
+        // Set the jump collision for server-executed entities that may control the boat
         bbmBoat.setJumpCollision(true);
 
         // Set the jump collision for players controlling the boat