+++ src/main/kotlin/dev/nyon/bbm/config/screen/ModMenuImpl.kt
@@ -1,7 +1,7 @@
 package dev.nyon.bbm.config.screen
 
 /*? if fabric {*/
-import com.terraformersmc.modmenu.api.ConfigScreenFactory
+/*import com.terraformersmc.modmenu.api.ConfigScreenFactory
 import com.terraformersmc.modmenu.api.ModMenuApi
 
 @Suppress("unused")
@@ -12,4 +12,4 @@ class ModMenuImpl : ModMenuApi {
         }
     }
 }
-/*?}*/
+*//*?}*/