+++ src/main/kotlin/dev/nyon/bbm/logic/BbmBoat.kt
@@ -2,4 +2,5 @@ package dev.nyon.bbm.logic
 
 interface BbmBoat {
     var jumpCollision: Boolean
+    var correctCollision: Boolean
 }